refactor string mutation operators
